recipemenu/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.7.0
django-filter==2.4.0
django-webpack-loader==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
